Check grounding for `rule` method
