+++ src/main/kotlin/io/github/alexandrepiveteau/datalog/core/interpreter/DatalogProgramBuilder.kt
@@ -1,10 +1,7 @@
 package io.github.alexandrepiveteau.datalog.core.interpreter
 
 import io.github.alexandrepiveteau.datalog.core.*
-import io.github.alexandrepiveteau.datalog.core.rule.Atom
-import io.github.alexandrepiveteau.datalog.core.rule.Predicate
-import io.github.alexandrepiveteau.datalog.core.rule.Rule
-import io.github.alexandrepiveteau.datalog.core.rule.Variable
+import io.github.alexandrepiveteau.datalog.core.rule.*
 
 /**
  * An implementation of [ProgramBuilder] which can be used to build a [Program].
@@ -26,11 +23,34 @@ internal class DatalogProgramBuilder<T>(
 
   private val rules = mutableSetOf<Rule<T>>()
 
+  private fun Rule<T>.limited(): Set<Variable> {
+    val variables = mutableSetOf<Variable>()
+    body
+        .asSequence()
+        .filter { !it.negated }
+        .forEach { clause -> variables.addAll(clause.atoms.filterIsInstance<Variable>()) }
+    when (this) {
+      is CombinationRule -> Unit
+      is AggregationRule -> variables.add(this.aggregate.result)
+    }
+    return variables
+  }
+
+  private fun requireGrounding(rule: Rule<T>) {
+    val head = rule.head.atoms.filterIsInstance<Variable>()
+    val body = rule.body.flatMap { it.atoms }.filterIsInstance<Variable>()
+    val limited = rule.limited()
+    for (variable in head + body) {
+      if (variable !in limited) throw NotGroundedException()
+    }
+  }
+
   override fun rule(predicate: Predicate, atoms: List<Atom<T>>, block: RuleBuilder<T>.() -> Unit) {
-    rules.add(DatalogRuleBuilder<T>().apply(block).toRule(predicate, atoms))
+    rule(DatalogRuleBuilder<T>().apply(block).toRule(predicate, atoms))
   }
 
   override fun rule(rule: Rule<T>) {
+    requireGrounding(rule)
     rules.add(rule)
   }
